pyproject.toml
[tool.poetry]
name = "buglog"
version = "0.1.0"
description = "Log what is happening to you"
authors = ["Lain Iwakura <lainiwa@cock.li>"]

[tool.poetry.dependencies]
python = "^3.7"
returns = "^0.13.0"
pyxdg = "^0.26"
importlib_metadata = "^1.6.0"
pydantic = {version = "^1.4", extras = ["typing_extensions"]}
typer = {version = "^0.1.1", extras = ["all"]}
pandas = "^1.0.3"

[tool.poetry.dev-dependencies]
pudb = "^2019.2"
ipython = "^7.13.0"
mypy = "^0.770"
pytest = "^5.4.1"
pre-commit = "^2.2.0"
pylint = "^2.4.4"
vulture = "^1.4"
pprintpp = "^0.4.0"

[tool.poetry.scripts]
buglog = 'buglog.console:app'

[tool.isort]
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 79
atomic = true
include_trailing_comma = true
indent = '    '
multi_line_output = 3
length_sort = true

[tool.black]
# pre-commit run black --help
line-length = 79
verbose = true
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
